Update Game.rkt
defined a lobby scene (lobbyScene) using the pre-defined lobbyWidth and lobbyHeight functions.

defined a variable called POINTER for the TEMPORARY starting point of the pointer.

changed the parameter name from image to img for the pointer structure.

deleted the useless function drawPointer after defining it due to temporary stupidness and instead abdulrahman combined it inside the menu function.